Finish consolidation of the module code
