+++ opencog/cogserver/modules/commands/BuiltinRequestsModule.cc
@@ -22,12 +22,10 @@ DECLARE_MODULE(BuiltinRequestsModule)
 
 BuiltinRequestsModule::BuiltinRequestsModule(CogServer& cs) : Module(cs)
 {
-/*
     _cogserver.registerRequest(ShutdownRequest::info().id,     &shutdownFactory);
     _cogserver.registerRequest(LoadModuleRequest::info().id,   &loadmoduleFactory);
     _cogserver.registerRequest(UnloadModuleRequest::info().id, &unloadmoduleFactory);
     _cogserver.registerRequest(ListModulesRequest::info().id,  &listmodulesFactory);
-*/
 
     do_help_register();
     do_h_register();
@@ -55,12 +53,10 @@ BuiltinRequestsModule::~BuiltinRequestsModule()
 
     do_stats_unregister();
 
-/*
     _cogserver.unregisterRequest(ShutdownRequest::info().id);
     _cogserver.unregisterRequest(LoadModuleRequest::info().id);
     _cogserver.unregisterRequest(UnloadModuleRequest::info().id);
     _cogserver.unregisterRequest(ListModulesRequest::info().id);
-*/
 }
 
 void BuiltinRequestsModule::init()
