+++ opencog/cogserver/modules/commands/ModuleManagement.cc
@@ -9,27 +9,7 @@
 #include <sstream>
 #include <string>
 
-#include <opencog/cogserver/server/CogServer.h>
-#include <opencog/cogserver/server/Request.h>
-#include <opencog/cogserver/server/RequestClassInfo.h>
-
-#define DEFINE_REQUEST(REQUESTNAME)                                   \
-                                                                      \
-namespace opencog                                                     \
-{                                                                     \
-class REQUESTNAME : public Request {                                  \
-public:                                                               \
-    REQUESTNAME(CogServer& cs) : Request(cs) {};                      \
-    virtual ~REQUESTNAME() {};                                        \
-    static const RequestClassInfo& info(void);                        \
-    virtual bool execute(void);                                       \
-    virtual bool isShell(void) {return info().is_shell;}              \
-}; }
-
-
-DEFINE_REQUEST(ListModulesRequest)
-DEFINE_REQUEST(LoadModuleRequest)
-DEFINE_REQUEST(UnloadModuleRequest)
+#include <opencog/cogserver/modules/commands/ModuleManagement.h>
 
 using namespace opencog;
